Split `grain_dump` into `grain_compile` and `grain_dump`.
